update: change floating window background colour
The colour cp.gray._1000 does not yet exist, so the background colour of
floating windows defaulted to the same colour as normal windows. This
looks confusing when a floating window overlays a normal window.

The colour has been changed to `#0C121D`, and is now visible when normal
and floating windows overlap.