feat: add `:Tundra` command
The command `:Tundra` is effectively a wrapper around the `colorscheme`
argument provided in Neovim. It will change the value of the global
variable `tundra_biome` and then call `:colorscheme tundra`.

The command takes one argument which is either the name of the biome (in
this case, `alpine` or `arctic`) or a background colour (`light` or
`dark`) and will change the highlight groups according to the argument
added by the user.

The first conditional check in the function ensures that the function
fails on any key or value that is NOT found in the `biomes` table in
`biomes.lua`.