feat: allow user configuration to overwrite syntax groups
This commit introduces a way to overwrite highlight groups from the
setup function, in a simple way similar to that of Catppuccin and
Nightfox.

Though it is meant to be used for conventionally adding
`bold` or `italic` options to syntax groups, it can also be used to
overwrite existing colors by `fg` and `bg` options, as well as reversing
groups with the `reverse` option.