fix: allow theme to be set without explicit `setup()` call
Before, if the global variable `tundra_opts` did not exist (because the
users' Neovim configuration file did not call the tundra `setup()`
function) it would set `local opts = nil`, causing an error when
`opts.plugins` was looped on later in the `load()` function.

This commit ensures that if the user does not call the `setup()`
function, the default configuration will still be respected.