fix: `deep_extend` overwrite configurations
A bug was introduced in commit 84e5d38 where adding a colour using the
user `overwrite` table caused the whole colour table to be overwritten.
This meant that if the colour contained different shades, and they were
not specified in the overwrite table, these colours would be
overwritten. Deep extending the table allows only the specific shade to
be added/overwritten, rather than the whole colour table.