fix: remove global configuration dependency
Bug fix that allows the colorscheme to work as normal if no `setup()`
function is defined by the user. Before this, if sub-tables inside the
global variable `tundra_opts` were not set it would break the code. This
commit adds a conditional around this global variable, and sets
`overwrite` to an empty table if any sub-tables are not found.