fix: correctly reload on `:colorscheme tundra` command
This commit fixes an issue when loading another colorscheme at Neovim
start-up, and then changing to the Tundra colorscheme using the command
`:colorscheme tundra`. Highlight groups for plugins were not set in the
example above, and this was due to the synchronous nature of the
`load()` function in `init.lua`.

This was because plugin highlight groups were being set before any base
highlights (built-in syntax and editor highlights). As plugin groups
were linked to these base groups using the `link` functionality in
`nvim_set_hl()` it meant that they weren't being linked to Tundra's
highlight groups, but instead the groups of the first colorscheme that
was called.

Swapping the location of plugin loading so that they are called *after*
base highlights are set in the `load()` function fixes this issue.