update: add completion to `:Tundra` command
This command is redundant at the moment as there is only one theme for
Tundra. In the future, I'd like this tundra command to be used for
changing user configuration options 'on the fly', so that a user does
not need to reload their Neovim configuration to change settings such as
`dim_inactive_windows.enabled`.